+++ resources/config_global.rb
@@ -1,7 +1,7 @@
 property :haproxy_user, String, default: 'haproxy'
 property :haproxy_group, String, default: 'haproxy'
 property :pidfile, String, default: '/var/run/haproxy.pid'
-property :log, String, default: '/dev/log syslog info'
+property :log, [String, Array], default: '/dev/log syslog info'
 property :daemon, [TrueClass, FalseClass], default: true
 property :debug_option, String, default: 'quiet', equal_to: %w(quiet debug)
 property :stats, Hash, default: lazy {
@@ -36,8 +36,8 @@ action :create do
       variables['global']['group'] = new_resource.haproxy_group
       variables['global']['pidfile'] ||= ''
       variables['global']['pidfile'] << new_resource.pidfile
-      variables['global']['log'] ||= ''
-      variables['global']['log'] = new_resource.log
+      variables['global']['log'] ||= []
+      variables['global']['log'] << new_resource.log
       variables['global']['log_tag'] ||= ''
       variables['global']['log_tag'] << new_resource.log_tag
       variables['global']['chroot'] ||= '' unless new_resource.chroot.nil?