+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 ## [unreleased]
 - updating service to use cookbook template
+- Add option for install only #251
+- `log` `property` in `global` resource can now be of type `Array` or `String`. This fixes #252
 
 ## [v4.6.1] (2017-08-02)
 - Reload instead of restart on config change